Changes for white bg-color
